Add types for outputs
