+++ cli/src/main.rs
@@ -1,8 +1,6 @@
 use clap::Parser;
-use commands::*;
+use commands::{sidekick::PreExecuteHandled, *};
 use global_error::prelude::*;
-use serde_json::json;
-use util::struct_fmt::{self, Format};
 
 use util::{global_config, os, term};
 
@@ -180,44 +178,9 @@ async fn main_inner(opts: Opts) -> GlobalResult<()> {
 
 	// Sidekick sign-in can also be called before the token is valitdated
 	if let SubCommand::Sidekick { command } = &opts.command {
-		let response: GlobalResult<_> = match command {
-			sidekick::SubCommand::GetLink { .. } => command.get_link().await,
-			sidekick::SubCommand::WaitForLogin { device_link_token } => {
-				command.wait_for_login(device_link_token).await
-			}
-			sidekick::SubCommand::CheckLoginState => command.validate_token(&token),
-			_ => {
-				// If the command is anything else, we need to check if a token
-				// has already been provided. If not, we need to print an error
-				// and return early since that's what the plugins will expect.
-				if let Err(_) = command.validate_token(&token) {
-					// The message has already been printed out so we can just
-					// return Ok here.
-					Ok(sidekick::SideKickResponse(json!({
-						"output": "Token not found. Please run `rivet sidekick get-link` to sign in."
-					})))
-				} else {
-					Ok(sidekick::SideKickResponse(json!({})))
-				}
-			}
-		};
-
-		// Print the response
-		match response {
-			Ok(sidekick_response) => {
-				struct_fmt::print(&Format::Json, &json!({ "Ok": sidekick_response }))?;
-			}
-			Err(global_error) => {
-				struct_fmt::print(
-					&Format::Json,
-					&json!({
-						"Err": global_error.to_string()
-					}),
-				)?;
-			}
+		if let Ok(PreExecuteHandled::Yes) = command.pre_execute(&token).await {
+			return Ok(());
 		}
-
-		return Ok(());
 	}
 
 	let Some(token) = token else {
@@ -257,19 +220,7 @@ async fn main_inner(opts: Opts) -> GlobalResult<()> {
 		SubCommand::Engine { command } => command.execute(&ctx).await?,
 		SubCommand::Unreal { command } => command.execute(&ctx).await?,
 		SubCommand::CI { command } => command.execute(&ctx).await?,
-		SubCommand::Sidekick { command } => match command.execute(&ctx, &term).await {
-			Ok(sidekick_response) => {
-				struct_fmt::print(&Format::Json, &json!({ "Ok": sidekick_response }))?;
-			}
-			Err(global_error) => {
-				struct_fmt::print(
-					&Format::Json,
-					&json!({
-						"Err": global_error.to_string()
-					}),
-				)?;
-			}
-		},
+		SubCommand::Sidekick { command } => command.execute(&ctx, &term).await?,
 	}
 
 	Ok(())
